Added error handling for FDW read flow (#900)
* ported json error handling and other logic from ext table code

* added custom error callback to FDW, error token in server

* updated avro test, annotations, global init file

* use define macros instead of constants

* rebased from main

* fixed Json test

* added building / installing standalone PXF FDW extension for GP6 / RHEL8

* fixed script

* rewired resource, made sure /tmp is writable for everyone

* list /tmp directory

* test FDW, dynamic profiles, filter pushdown test

* fixed new line, chmod

* use round to reduce precision

* processing of reject limits in tests

* adjusted test matchsubs

* adjusted match subs

* fixed test match/sub for GP5

* updated ColumnProjectionTest

* check for foreign server for GP6

* removed TEST profile logic from HttpRequestParser

* addressed PR review comments

* fix script and matchsubs comment

* PR review comment

* adjusted expected error message in test

* addressed PR feedback

* adjusted regex for error table test

* adjusted regex to non-greedy for error table test

* added dynamic profile property for multi-node setup

* fix install of pxf-fdw-* tarball for PR pipeline

+++ automation/src/test/java/org/greenplum/pxf/automation/features/cloud/CloudAccessTest.java
@@ -62,8 +62,10 @@ public class CloudAccessTest extends BaseFeature {
     @Override
     protected void afterMethod() throws Exception {
         super.afterMethod();
-        s3Server.removeDirectory(PROTOCOL_S3 + s3PathRead);
-        s3Server.removeDirectory(PROTOCOL_S3 + s3PathWrite);
+        if (s3Server != null) {
+            s3Server.removeDirectory(PROTOCOL_S3 + s3PathRead);
+            s3Server.removeDirectory(PROTOCOL_S3 + s3PathWrite);
+        }
     }
 
     protected void prepareData() throws Exception {
